fix: revert minimized calling cell
