+++ src/script/components/Conversation/Conversation.tsx
@@ -490,16 +490,17 @@ export const Conversation = ({
             }
 
             return (
-              <CallingCell
-                key={conversation.id}
-                classifiedDomains={classifiedDomains}
-                call={call}
-                callActions={callingViewModel.callActions}
-                callingRepository={callingRepository}
-                pushToTalkKey={callingViewModel.propertiesRepository.getPreference(
-                  PROPERTIES_TYPE.CALL.PUSH_TO_TALK_KEY,
-                )}
-              />
+              <div className="calling-cell" key={conversation.id}>
+                <CallingCell
+                  classifiedDomains={classifiedDomains}
+                  call={call}
+                  callActions={callingViewModel.callActions}
+                  callingRepository={callingRepository}
+                  pushToTalkKey={callingViewModel.propertiesRepository.getPreference(
+                    PROPERTIES_TYPE.CALL.PUSH_TO_TALK_KEY,
+                  )}
+                />
+              </div>
             );
           })}
