+++ src/script/components/calling/CallingCell.test.tsx
@@ -72,6 +72,7 @@ const createProps = async () => {
     pushToTalkKey: null,
     conversation,
     hasAccessToCamera: true,
+    isSelfVerified: true,
     teamState: mockTeamState,
     videoGrid: {grid: [], thumbnail: undefined},
   } as CallingCellProps;
