+++ src/style/components/group-video-grid.less
@@ -40,7 +40,7 @@
     overflow: hidden;
     height: 168px;
     border-radius: 8px;
-    margin-bottom: 8px;
+    margin: 8px;
 
     &--no-camera-access {
       display: flex;
