+++ src/style/list/conversation-list-cell.less
@@ -71,6 +71,12 @@
   }
 }
 
+.conversation-list-cell-button {
+  overflow: hidden;
+  width: 100%;
+  flex: 1;
+}
+
 .conversation-list-cell-left {
   .flex-center;
 
@@ -96,6 +102,23 @@
 
   padding-right: 16px;
   border-bottom: 1px solid var(--border-color);
+
+  &__calling {
+    position: relative;
+    display: flex;
+    width: 100%;
+    height: 56px;
+    border-radius: 8px 8px 0 0;
+    cursor: pointer;
+    font-weight: @font-weight-regular;
+    & > * {
+      border-radius: 8px;
+    }
+    & > *,
+    & > * > * {
+      border: none;
+    }
+  }
 }
 
 .conversation-recent-view-list {
